refactor: cli apllication
