+++ src/cli/commands/index.ts
@@ -1,0 +1,2 @@
+export * from './import.js';
+export * from './generate.js';
